requirements.txt
click==6.7
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
flask-bower==1.3.0
azure-storage-file-datalake==12.0.0b6
numpy==1.16.6
opencv-python==4.1.2.30
matplotlib==3.1.2